// src/types/responceTypes.ts
export interface SearchResult {
  from: number
  to: number
  count: number
  _links: Links
  hits: Hit[]
}

export interface Links {
  next: Next
}

export interface Next {
  href: string
  title: string
}

export interface Hit {
  recipe: Recipe
  _links: Links2
}

export interface Recipe {
  uri: string
  label: string
  image: string
  images: Images
  source: string
  url: string
  shareAs: string
  yield: number
  dietLabels: string[]
  healthLabels: string[]
  cautions: string[]
  ingredientLines: string[]
  ingredients: Ingredient[]
  calories: number
  totalWeight: number
  totalTime: number
  cuisineType: string[]
  mealType: string[]
  dishType: string[]
  totalNutrients: TotalNutrients
  totalDaily: TotalDaily
  digest: Digest[]
}
// ...
export interface Links2 {
  self: Self
}